Camera update - allow non-MJPEG as fullscreen
Update:
When the videoUrl is not populated, the camera block will use the imageUrl. This is for cameras with MJPEG capability, or where the user cannot find one. The fullscreen video will be replace with image refreshing, like on the block preview. All embedded HTML has been exported to templates.